+++ libs/core/src/lib/info-label/info-label.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { InfoLabelComponent } from './info-label.component';
+
+@NgModule({
+    declarations: [InfoLabelComponent],
+    imports: [
+        CommonModule
+    ],
+    exports: [InfoLabelComponent]
+})
+export class InfoLabelModule { }