+++ libs/core/src/lib/utils/functions/index.ts
@@ -1,1 +1,3 @@
-export {compareObjects} from './compare-objects'
+export { compareObjects } from './compare-objects';
+export { getStringFromHashMap } from './get-string-from-hashmap';
+export { uuidv4 } from './uuidv4-generator';